@@ -43,13 +43,13 @@ dev = [
 
 [tool.black]
 line-length = 100
-target-version = ["py310"]
+target-version = ["py311"]
 
 [tool.ruff]
 fix = true
 line-length = 100
 select = ["B", "E", "F", "I"]
-target-version = "py310"
+target-version = "py311"
 
 [tool.ruff.format]
 docstring-code-format = true
@@ -58,7 +58,7 @@ docstring-code-format = true
 check_untyped_defs = true
 disallow_untyped_defs = true
 ignore_missing_imports = true
-python_version = "3.10"
+python_version = "3.11"
 show_error_codes = true
 strict = true
 warn_return_any = true
