@@ -24,7 +24,7 @@ dependencies = [
     "typing-extensions>=4.9.0",
     "pydantic>=2.6.3"
 ]
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 readme = "README.md"
 license = {text = "MIT"}
 keywords = ["computer-vision", "ocr", "ui-analysis", "icon-detection"]
